backend/requirements.txt
# Backend Dependencies for Bloom (Flask API)
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
SQLAlchemy>=2.0
gunicorn==21.2.0
libsql-client==0.3.1
psycopg2-binary==2.9.9

# Data validation and serialization
marshmallow==3.20.1
email-validator==2.1.0

# Email service
sendgrid==6.11.0

# Date/time handling
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
